agent/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
kafka-python==2.0.2
clickhouse-driver==0.2.6
